requirements.txt
flask
face-recognition
faiss-gpu
opencv-python
azure-ai-vision-imageanalysis==1.0.0b1
requests
fuzzywuzzy
python-Levenshtein
pandas

# anime
torchvision
transformers
sentence_transformers

# logo
omegaconf
torchvision
tensorflow==2.15.0
keras
scipy
scikit-learn

# numpy==1.23.0
# # super-gradients==3.5.0
# # opencv-python-headless
# # tensorflow==2.15.0
# # keras-vggface==0.6
# keras==2.15.0
# scikit-learn==1.3.2
# joblib==1.3.2
matplotlib
                                                                                                                                                                                                